Rename package to noteable
