+++ noteable/util.py
